rename loader utils
